+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* NP-239 Advice Feedback
+
 ## <sub>v0.11.0</sub>
 
 #### _May. 5, 2020_